update
integrating jqgrid completely into lcms.js
