+++ src/main/webapp/HTML/pages/template/index.html
@@ -7,10 +7,7 @@
         <script src="./HTML/pages/template/interface.js"></script>
         <link rel="stylesheet" type="text/css" media="screen" href="./HTML/pages/template/style.css" />
     </head>
-    <body>
-       
-
-        <div>LCMSEditablePage-menu</div>
+    <body>      
         <div id="sidebar"></div>
         <div class="container" id="page-container">
 
@@ -42,13 +39,4 @@
     </body>
 </html>
 
-<script>
-//    $("div[contenteditable]").each(function (a, b) {
-//        var ck = CKEDITOR.inline($(b).attr('id'));
-//        ck.on('instanceReady', function (ev) {
-//            var editor = ev.editor;
-//            editor.setReadOnly(false);
-//        });
-//        ;
-//    });
-</script>
+
