+++ src/main/webapp/HTML/pages/template/menu.html
@@ -1,12 +1,11 @@
 <!DOCTYPE html>
 
-<div id="div-page-menu">
-    <button type="button" id='btn-page-edit' class="btn btn-info" onclick="edit_page()"><i class="fa fa-lg fa-fw fa-edit"></i><span>Bewerken</span></button>
+<!--<div id="div-page-menu">
+    <button type="button" id='btn-page-edit' class="btn btn-info" onclick="documentPage.edit_page()"><i class="fa fa-lg fa-fw fa-edit"></i><span>Bewerken</span></button>
     <div id="edit-menu">
         <button type="button" id='btn-page-save' class="btn btn-primary" onclick="documentPage.savePage()" disabled><i class="fa fa-lg fa-fw fa-save"></i><span>Opslaan</span></button>
-        <button type="button" class="btn btn-primary" data-toggle="collapse" data-target="#collapseInvoegen" role="button" aria-expanded="false" aria-controls="collapseInvoegen" disabled>
-            Invoegen
-        </button> <div class="collapse" id="collapseInvoegen">
+        <button type="button" class="btn btn-primary" data-toggle="collapse" data-target="#collapseInvoegen" role="button" aria-expanded="false" aria-controls="collapseInvoegen" disabled>Invoegen</button>
+        <div class="collapse" id="collapseInvoegen">
             <div class="card card-body">
                 <button type="button" id='btn-new-editable-field' class="btn btn-primary" onclick="new_editable_field()" disabled><i class="fa fa-lg fa-fw fa-align-justify"></i><span>Tekstveld</span></button>
                 <button type="button" id='btn-new-jqgrid' class="btn btn-primary" onclick="new_grid('page-container')" disabled><i class="fa fa-lg fa-fw fa-table"></i><span>Tabel</span></button>
@@ -14,13 +13,11 @@
 
             </div>
         </div>
-        <button type="button" class="btn btn-primary" data-toggle="collapse" data-target="#collapseActies" role="button" aria-expanded="false" aria-controls="collapseActies" disabled>
-            Acties
-        </button> <div class="collapse" id="collapseActies">
+        <button type="button" class="btn btn-primary" data-toggle="collapse" data-target="#collapseActies" role="button" aria-expanded="false" aria-controls="collapseActies" disabled>Acties</button>
+        
+        <div class="collapse" id="collapseActies">
             <div class="card card-body">
-                <label class="btn btn-primary btn-file">
-                    <i class="fa fa-lg fa-fw fa-th"></i><span>Importeer</span> <input id="importTableButton" type="file" style="display: none;">
-                </label>    
+                <label class="btn btn-primary btn-file"><i class="fa fa-lg fa-fw fa-th"></i><span>Importeer</span> <input id="importTableButton" type="file" style="display: none;"></label>    
                 <button type="button" id='btn-export-html' class="btn btn-primary" onclick="exportToHTML()" disabled><i class="fa fa-lg fa-fw fa-download" ></i><span>Download</span></button>
             </div>
         </div>
@@ -30,4 +27,4 @@
         <table id="history-table"></table>
         <div id="history-pager"></div>
     </div>
-</div>
+</div>-->
